paper_automation/pyproject.toml
[project]
name = "paper-automation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.12.15",
    "bs4>=0.0.2",
    "google>=3.0.0",
    "google-generativeai>=0.8.5",
    "lxml>=6.0.0",
    "openpyxl>=3.1.5",
    "pandas>=2.3.1",
    "playwright>=1.54.0",
    "pyside6>=6.9.1",
    "pytest-playwright>=0.7.0",
    "python-dotenv>=1.1.1",
]
